+++ packages/services/src/test-utils.ts
@@ -33,9 +33,10 @@ export interface IndexedDbMock {
   getPricesForAsset?: Mock;
   getAuction?: Mock;
   getAuctionOutstandingReserves?: Mock;
-  hasStakingAssetBalance?: Mock;
   stakingTokenAssetId?: Mock;
   upsertAuction?: Mock;
+  hasTokenBalance?: Mock;
+  saveGasPrices?: Mock;
 }
 
 export interface AuctionMock {